Implemented separate classifier layer.
Depending on whether args.share_item_emb is set to true or not,
we'll have a separate linear layer that acts as a classifier layer.

I'll also have to implement a if-else branch in the forward method.
I'm also curious if there's a way to simply use one classifier layer
and copy the weights. Would this be a deep copy? If so, that wouldn't
work.