create get_negative_labels + fix collate_fn
Fixed the collate_fn so that we're getting negative samples for
every positive sample. The official paper and code seem to say that
the authors created one negative sample for each positive sample
during training, and during evaluation they got 100 negative samples.

I changed the name of collate_fn to collate_fn_train because I have
a feeling that we're going to be needing a new collation stategy in
order to bypass the fact that creating negative samples up to 100
takes too long (around 15 seconds for one positive label).

I'm planning to create negative samples during the data preprocessing
stage for validation and test sets. It's better for the sampling to
take long in the beginning rather than go through that for every
single iteration.