Moved get_negative_samples and added pad_or_truncate_seq.
Rather than having the get_negative_samples as a method for the
dataset object, I thought that it'd make more sense to have it as
a separate function because I'll be calling it in other places too.

The pad_or_truncate_seq basically took all of the padding and truncating
operations in the collate functions and performed them outside. This
should not only enhance readability but also improve performance.