Made new calculate loss method and finished train.
Rather than have all those operations regarding indexing and whatnot
done in the main train loop, I made a separate calculate_bce_loss
method to perform all of that.

The train method loop is essentially finished now and I just have to
test it and make sure the loss decreases.

[What next?]
Need to implement the evaluation function and add that inside the
training loop.