Changed code for simpler weight sharing.
The previous code had a if-else block inside of the forward pass that
did not look that pretty. I changed the code so that rather than
having a separate block inside the forward pass, we simply assign the
weights of the item embedding matrix to the classifier if the option
to share weights is turned on.