Added collate_fn_eval and modified get_dataloader.
The current DataLoader was only suitable for the training phase.
During the evaluation phase we need both the input sequence and
the positive sample with the 100 negative samples.

collate_fn_eval essentially returns both of those in a tuple format
so that we can simply use the DataLoaders during evaluation.