Added self.optimizer.step() and moved zero_grad().
I didn't realize I had left out the self.optimizer.step() line and
was getting a warning that I was calling self.scheduler.step() before
that.

I also moved self.model.zero_grad() to the start of the for loop so
that we don't have to worry about it in subsequent stages.