Implemented prediction in forward method.
The forward method is now able to handle the validation and test
splits. In the training phase, we pass the positive and negative
sequences to forward, but in the evaluation phase(s) we pass through
the items that we want to rank.

Depending on which is passed, we'll get the logits.

[To-Do]
Q: Do we need the classifier layer at all?