Added block to get resumption dir.
When we're resuming training, we either have the option of explicitly
telling the script which checkpoint to use or not.

If the resume_dir is provided, then we load our args, models, and log
files from there. If not, then the most recent one among the given data
name is used.